pyproject.toml
[tool.poetry]
name = "missing_persons"
version = "0.1.0"
description = "A repository of missing persons from police abductions"
authors = ["Teddy Ngeene <tngeene@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-filter = "^24.2"
pillow = "^10.3.0"
django-allauth = "^0.62.1"
django-bootstrap-datepicker-plus = "^5.0.5"
django-environ = "^0.11.2"
django-model-utils = "^4.5.1"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.5.1"
django-extensions = "^3.2.3"
django-phonenumber-field = {version = "^7.3.0", extras = ["phonenumberslite"]}
django-simple-history = "^3.5.0"
django-widget-tweaks = "^1.5.0"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
ipython = "^8.24.0"
django-bootstrap4 = "^24.3"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
